requirements.txt
astroid==2.2.5
autopep8==1.4.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-heroku==0.3.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
python-decouple==3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.3
wrapt==1.11.2
